PTU-PAYGO-OpenAI-Simulator/PTU-PAYGO-OpenAI-Simulator/OPTIMIZATION_ANALYSIS.py
"""
OPTIMIZATION ANALYSIS: Azure Log Parallel Processing
=====================================================

Current Performance (100 workers):
- Rate: 6.9 blobs/sec = 414 blobs/min = 24,840 blobs/hour
- Active connections: ~123 (should be closer to 100)
- CPU: 17-60% (lots of headroom)
- Memory: Plenty available

BOTTLENECK ANALYSIS:
===================

1. **Azure Authentication Overhead** ⚠️ MAJOR
   - Each worker uses DefaultAzureCredential()
   - This spawns Azure CLI subprocesses for token refresh
   - We saw 100+ "az account get-access-token" processes
   - Token refresh happens periodically, blocking workers
   
   Solution: Share credentials across workers
   - Use a single token manager
   - Pre-authenticate before spawning pool
   - Pass token/credential to workers instead of recreating

2. **Connection Pool Size** ⚠️ MODERATE
   - Each worker creates its own BlobServiceClient
   - Azure SDK may be limiting concurrent connections
   - Default HTTP connection pool might be conservative
   
   Solution: 
   - Increase max_connections in BlobServiceClient
   - Use connection pooling across workers
   - Configure retry policy more aggressively

3. **Batch Size** ⚠️ MINOR
   - Processing in chunks of 100 blobs
   - Could be larger for better throughput
   
   Solution: Increase to 500-1000 blob chunks

4. **JSON Parsing** ⚠️ MINOR
   - NDJSON parsing is CPU-bound but fast
   - Not the limiting factor
   
   Already optimized.

5. **CSV Writing** ⚠️ MINIMAL
   - Writing at end, not during processing
   - Not a bottleneck

OPTIMIZATION STRATEGIES:
========================

## Strategy 1: Shared Authentication (BEST - 2-3x speedup)
```python
# Pre-authenticate once
credential = DefaultAzureCredential()
# Get token
token = credential.get_token("https://storage.azure.com/.default")

# Pass token to workers instead of credential
# Workers use token directly without subprocess calls
```

## Strategy 2: Increase Workers (EASY - 1.5-2x speedup)
```bash
# Current: 100 workers
# Recommended: 200-300 workers
# Your Mac can handle it (I/O bound, not CPU bound)

python3 download_azure_logs_parallel.py --workers 200 --force
```

## Strategy 3: Connection Pool Tuning (MODERATE - 1.3-1.5x speedup)
```python
from azure.storage.blob import BlobServiceClient
from azure.core.pipeline.transport import RequestsTransport

# Increase connection pool
transport = RequestsTransport(
    connection_pool_maxsize=500,
    connection_pool_block=False
)

blob_service = BlobServiceClient(
    account_url=account_url,
    credential=credential,
    transport=transport
)
```

## Strategy 4: Async/Await (COMPLEX - 3-5x speedup)
```python
# Replace multiprocessing with asyncio
# Use aiohttp for concurrent downloads
# Can handle 1000+ concurrent requests easily

async with BlobServiceClient(...) as client:
    tasks = [download_blob(blob) for blob in blobs]
    await asyncio.gather(*tasks, return_exceptions=True)
```

RECOMMENDED APPROACH FOR NEXT TIME:
===================================

**Phase 1: Quick Wins (30 min implementation)**
1. Increase workers to 200-300
2. Increase batch size to 500
3. Add connection pool tuning
Expected: 2-3x speedup (1.5-2 hours instead of 4.5)

**Phase 2: Major Optimization (2-3 hour implementation)**
1. Implement shared authentication token
2. Use asyncio instead of multiprocessing
3. Batch downloads with connection reuse
Expected: 5-10x speedup (30-60 minutes instead of 4.5 hours!)

**Phase 3: Ultimate Performance (1 day implementation)**
1. Async processing with aiobotocore
2. Stream processing (don't store all in memory)
3. Progressive CSV writing
4. Multiple Azure storage accounts (parallel accounts)
Expected: 10-20x speedup (15-30 minutes for 120K blobs!)

PROOF OF CONCEPT CODE:
======================

See: download_azure_logs_optimized.py (to be created)
- Uses asyncio + aiohttp
- Shared credential manager
- Connection pooling
- Chunked CSV writing
- Progress streaming

HARDWARE HEADROOM:
==================

Your M4 Mac can handle MUCH more:
- CPU: Only using 17-60% (can handle 3-5x more workers)
- Memory: 13GB/16GB used (plenty of room)
- Network: Gigabit capable (not saturated)
- Disk: SSD can handle concurrent writes

The bottleneck is NOT your machine - it's:
1. Azure authentication overhead
2. SDK connection management
3. Serial processing in each worker

CONSERVATIVE ESTIMATE FOR NEXT TIME:
====================================

With just Strategy 1 + 2 (simple changes):
- Workers: 200-300
- Shared auth: No subprocess overhead
- Expected rate: 15-20 blobs/sec (vs current 6.9)
- Expected time: 1.5-2 hours (vs current 4.5)
- Implementation: 30 minutes

AGGRESSIVE ESTIMATE:
===================

With full async implementation:
- Concurrent connections: 500-1000
- Rate: 30-50 blobs/sec
- Time: 30-60 minutes
- Implementation: 2-3 hours (one-time)

CONCLUSION:
===========

YES - We can drive it MUCH harder! 🚀

The current implementation is conservative and safe.
With optimizations, we can achieve:
- 3x speedup with minimal code changes
- 10x speedup with async implementation
- Potentially 20x with ultimate optimization

Your Mac can handle it - the bottleneck is software architecture,
not hardware capabilities.

For production use, I recommend implementing Phase 1 + 2 
for the best balance of effort vs. performance gain.
"""

if __name__ == '__main__':
    print(__doc__)
