PTU-PAYGO-OpenAI-Simulator/PTU-PAYGO-OpenAI-Simulator/pyproject.toml
[project]
name = "ptu-calc"
version = "0.1.0"
description = "PTU vs PAYGO Traffic Simulator for Azure OpenAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "streamlit>=1.49.1",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]
